Implement vanity SSH key generator
Uses the `ssh-key` crate for all cryptography, looks to be an efficient
and secure implementation for managing SSH keys. Aside from that, just
plumbs together a secure OS CSPRNG with the SSH key fingerprinting and
regex matching into a multi-threaded searcher with rate reporting.

In the future, could consider validation of the regex for base64 and/or
crude estimates of time remaining based on regex match length and
current rate.